requirements.txt
Flask==2.0.2
google-generativeai==0.1.1